requirements.txt
fastapi[all]
uvicorn
python-dotenv
libsql-client
pymysql
dotenv
gunicorn 
sqlalchemy-libsql